+++ lib/phx_bb_web/live/components/users_online_component.ex
@@ -6,4 +6,6 @@ defmodule PhxBbWeb.UsersOnlineComponent do
   use PhxBbWeb, :live_component
 
   import PhxBbWeb.StyleHelpers
+
+  def guest?(id) when is_binary(id), do: String.at(id, 0) == "-"
 end